refactor: dao를 ChessGame에서 사용하도록 수정
