+++ src/main/java/dao/MoveDaoImpl.java
@@ -1,7 +1,9 @@
 package dao;
 
+import domain.Position;
 import domain.command.MoveCommand;
 import java.sql.Connection;
+import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -10,12 +12,29 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class MoveDaoImpl implements MoveDao {
+    private static final String SERVER = "localhost:13306"; // MySQL 서버 주소
+    private static final String DATABASE = "chess"; // MySQL DATABASE 이름
+    private static final String OPTION = "?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC";
+    private static final String USERNAME = "root"; //  MySQL 서버 아이디
+    private static final String PASSWORD = "root"; // MySQL 서버 비밀번호
+
     @Override
-    public void add(Connection connection, MoveCommand moveCommand) {
+    public Connection getConnection() {
+        try {
+            return DriverManager.getConnection("jdbc:mysql://" + SERVER + "/" + DATABASE + OPTION, USERNAME, PASSWORD);
+        } catch (final SQLException e) {
+            System.out.println("DB 연결 오류:" + e.getMessage());
+            e.printStackTrace();
+            return null;
+        }
+    }
+
+    @Override
+    public void add(Position from, Position to) {
         final String query = "INSERT INTO move (`from`, `to`) VALUES(?, ?)";
-        try (final PreparedStatement preparedStateMent = connection.prepareStatement(query)) {
-            preparedStateMent.setString(1, moveCommand.getFrom().name());
-            preparedStateMent.setString(2, moveCommand.getTo().name());
+        try (final PreparedStatement preparedStateMent = getConnection().prepareStatement(query)) {
+            preparedStateMent.setString(1, from.name());
+            preparedStateMent.setString(2, to.name());
             preparedStateMent.executeUpdate();
         } catch (final SQLException e) {
             throw new RuntimeException(e);
@@ -23,10 +42,10 @@ public class MoveDaoImpl implements MoveDao {
     }
 
     @Override
-    public List<MoveCommand> findAllMoves(Connection connection) {
+    public List<MoveCommand> findAllMoves() {
         final String query = "SELECT * FROM move";
         List<MoveCommand> moveCommands = new ArrayList<>();
-        try (final PreparedStatement preparedStatement = connection.prepareStatement(query)) {
+        try (final PreparedStatement preparedStatement = getConnection().prepareStatement(query)) {
             final ResultSet resultSet = preparedStatement.executeQuery();
             while (resultSet.next()) {
                 moveCommands.add(new MoveCommand(
@@ -41,9 +60,9 @@ public class MoveDaoImpl implements MoveDao {
     }
 
     @Override
-    public void deleteAll(Connection connection) {
+    public void deleteAll() {
         final String query = "TRUNCATE TABLE move";
-        try (final Statement statement = connection.createStatement()) {
+        try (final Statement statement = getConnection().createStatement()) {
             statement.executeUpdate(query);
         } catch (final SQLException e) {
             throw new RuntimeException(e);
